requirements.txt
discord.py
python-dotenv
beautifulsoup4
requests
html2text